Start split of Window.cpp into several smaller files
